feat/57: Refactor code
