+++ src/views/components/Classes/popupDetails.jsx
